Migrate role add functionality into separate component
